+++ lib/bitly.dart
@@ -1,7 +1,11 @@
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:flutter/foundation.dart';
+import 'package:flutter/widgets.dart';
+import 'package:flutter_url_shortener/bitly_exception.dart';
+import 'package:flutter_url_shortener/bitly_model.dart';
+import 'package:flutter_url_shortener/bitly_params.dart';
+import 'package:flutter_url_shortener/constant.dart';
 
 class FShort {
   FShort._internal();
@@ -22,24 +26,29 @@ class FShort {
 
   /// Generate shorten URL
   ///
-  /// @param [url] url that you want to shorten
+  /// @param [longUrl] url that you want to generate shorten
   ///
   /// Example
   /// ```
   /// try {
-  ///     FShort.instance
-  ///         .generateShortenURL(url: 'https://www.google.com.vn/')
-  ///         .then((value) {
-  ///             print("HCCCCCCC ${value}");
-  ///          });
-  ///     } catch(e) {
-  ///          print(e);
+  ///       FShort.instance
+  ///           .generateShortenURL(longUrl: 'https://www.google.com.vn/')
+  ///           .then((value) {
+  ///         setState(() {
+  ///           _shortenURL = value.link;
+  ///         });
+  ///       });
+  ///     } on BitlyException catch (_) {
+  ///
+  ///     } on Exception catch (_) {
+  ///
   ///     }
   /// ```
   ///
-  Future<String> generateShortenURL({
-    @required String url,
+  Future<BitlyModel> generateShortenURL({
+    @required String longUrl,
     String groupId,
+    String domain = 'bit.ly',
   }) async {
     final client = HttpClient();
     final endPoint = Uri.https('api-ssl.bitly.com', '/v4/shorten');
@@ -50,8 +59,8 @@ class FShort {
         ..set(HttpHeaders.authorizationHeader, 'Bearer $_token');
 
       final body = {
-        'long_url': url,
-        'domain': 'bit.ly',
+        'long_url': longUrl,
+        'domain': domain,
         'group_guid': groupId,
       };
       request.add(utf8.encode(json.encode(body)));
@@ -59,7 +68,66 @@ class FShort {
     });
 
     final responseBody = await response.transform(utf8.decoder).join();
-    final responseJson = json.decode(responseBody) as Map<String, dynamic>;
-    return responseJson['link'];
+
+    if (response.statusCode == StatusCode.SUCCESS ||
+        response.statusCode == StatusCode.CREATED) {
+      return BitlyModel.fromJson(json.decode(responseBody));
+    } else {
+      throw BitlyException.fromJson(json.decode(responseBody));
+    }
+  }
+
+  ///
+  /// Converts a long url to a Bitlink and sets additional parameters.
+  /// ```
+  /// try {
+  ///       FShort.instance
+  ///           .createBitLink(
+  ///               params: BitlyParams(
+  ///         longUrl: "https://dev.bitly.com",
+  ///         domain: 'bit.ly',
+  ///         tags: ['ver1.1', 'ver1.2'],
+  ///         deeplinks: [
+  ///           DeeplinkParams(
+  ///             appId: 'com.hades.test',
+  ///             appUriPath: '/store?id=123456',
+  ///             installUrl:
+  ///                 'https://play.google.com/store/apps/details?id=com.hades.test&hl=en_US',
+  ///             installType: 'promote_install',
+  ///           ),
+  ///         ],
+  ///       ))
+  ///           .then((value) {
+  ///         setState(() {
+  ///           _customURL = value.link;
+  ///         });
+  ///       });
+  ///     } on BitlyException catch (_) {
+  ///       //
+  ///     } on Exception catch (_) {
+  ///       //
+  ///     }
+  /// ```
+  ///
+  Future<BitlyModel> createBitLink({@required BitlyParams params}) async {
+    final client = HttpClient();
+    final endPoint = Uri.https('api-ssl.bitly.com', '/v4/bitlinks');
+
+    final response = await client.postUrl(endPoint).then((request) {
+      request.headers
+        ..set(HttpHeaders.contentTypeHeader, 'application/json')
+        ..set(HttpHeaders.authorizationHeader, 'Bearer $_token');
+
+      request.add(utf8.encode(json.encode(params.toJson())));
+      return request.close();
+    });
+
+    final responseBody = await response.transform(utf8.decoder).join();
+    if (response.statusCode == StatusCode.SUCCESS ||
+        response.statusCode == StatusCode.CREATED) {
+      return BitlyModel.fromJson(json.decode(responseBody));
+    } else {
+      throw BitlyException.fromJson(json.decode(responseBody));
+    }
   }
 }