+++ CHANGELOG.md
@@ -1,1 +1,10 @@
-## [1.0.0] - Bit.ly shortener
+## [1.0.0]
+
+Bit.ly shortener
+
+## [1.0.1]
+
+Support generate short link with 2 type
+
+* Converts a long url to a Bitlink.
+* Converts a long url to a Bitlink and sets additional parameters.